% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldipickr-package.R
\docType{package}
\name{maldipickr-package}
\alias{maldipickr}
\alias{maldipickr-package}
\title{maldipickr: Dereplicate And Cherry-pick Mass Spectrometry Spectra}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convenient wrapper functions for the analysis of matrix-assisted laser desorption/ionization-time-of-flight (MALDI-TOF) spectra data in order to select only representative spectra (aka cherry-pick). The package covers the preprocessing and dereplication steps needed to cluster MALDI-TOF spectra before the final cherry-picking step. This package relies heavily on the 'MALDIquant' package albeit with a 'tidyverse' flavour.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClavelLab/maldipickr}
  \item Report bugs at \url{https://github.com/ClavelLab/maldipickr/issues}
}

}
\author{
\strong{Maintainer}: Charlie Pauvert \email{cpauvert@users.noreply.github.com} (\href{https://orcid.org/0000-0001-9832-2507}{ORCID})

}
\keyword{internal}
