% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_spectra.R
\name{pick_spectra}
\alias{pick_spectra}
\title{Cherry-pick Bruker MALDI Biotyper spectra}
\usage{
pick_spectra(
  cluster_df,
  metadata_df = NULL,
  criteria_column = NULL,
  is_descending_order = TRUE,
  discard_regex = NULL,
  only_show_discarded = FALSE,
  is_sorted = FALSE
)
}
\arguments{
\item{cluster_df}{A tibble with clusters information
from the \link{similarity_to_clusters} or the \link{import_spede_clusters} function.}

\item{metadata_df}{Optional tibble with relevant metadata to guide the
picking process (e.g., OD600).}

\item{criteria_column}{Optional character indicating the column in \code{metadata_df}
to be used as a criteria.}

\item{is_descending_order}{Optional logical indicating whether to sort in
descending order or not.}

\item{discard_regex}{Optional \link{regex} to remove spectra and therefore
clusters from the picking decision (e.g., if they have been picked already).}

\item{only_show_discarded}{Optional logical to be used in conjunction to
\code{discard_regex} to \strong{only print} the names of the spectra to be discarded
as a verification step.}

\item{is_sorted}{Optional logical to indicate that the \code{cluster_df} is
already sorted by cluster based on (usually multiple) internal criteria to
pick the first of each cluster. This flag is \strong{overriden} if a \code{metadata_df}
is provided.}
}
\value{
A tibble with as many rows as \code{cluster_df} with an additional logical
column named \code{to_pick} to indicate whether the colony associated to the spectra
should be picked. If \code{metadata_df} is provided, then additional columns from
this tibble are added to the returned tibble. When the \code{discard_regex} is used,
the \code{to_pick} column of the tibble for the discarded spectra displays \code{FALSE}.
When \code{only_show_discarded} is used, it returns $0$ and print
out the list of spectra to be removed.
}
\description{
Using the clusters information, and potential additional metadata as external
criteria, spectra are labelled as to be picked for each cluster. Note that some
spectra and therefore clusters can be explicitly removed from the picking decision
using regular expressions, if they have been previously picked. If no metadata
are provided, the reference spectra of each cluster will be picked.
}
\examples{
# Load a toy example of a tibble of clusters created by the `similarity_to_clusters`
#  function.
clusters <- readRDS(system.file("clusters_tibble.RDS", package = "maldipickr"))

# By default and if no other metadata are provided, the function
#  picks the reference spectrum of each clusters.
pick_spectra(clusters)

# Simulate OD600 values with uniform distribution
#  for each of the colonies we measured with
#  the Bruker MALDI Biotyper
set.seed(104)
metadata <- dplyr::transmute(clusters, name = name, OD600 = runif(n = nrow(clusters)))
metadata

# Pick the spectra based on the highest OD600 value per cluster
pick_spectra(clusters, metadata, "OD600")

# Say that spectra 'species2_E11' was picked before in a previous experiment
#  we do not want to pick a cluster where species2_E11 is included to limit redundancy
# The function can peform a dry run to indicate which spectra would be removed
pick_spectra(clusters, discard_regex = "E11", only_show_discarded = TRUE)
# Once checked, the real dereplication can be performed.
pick_spectra(clusters, discard_regex = "E11")
}
