% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biotyper_report.R
\name{read_biotyper_report}
\alias{read_biotyper_report}
\title{Importing Bruker MALDI Biotyper CSV report}
\usage{
read_biotyper_report(path, best_hits = TRUE)
}
\arguments{
\item{path}{Path to the semi-colon separated table}

\item{best_hits}{A logical indicating whether to return only the best hit}
}
\value{
A tibble (see Details for the columns description)
}
\description{
The headerless table exported by the Compass software in the Bruker MALDI
Biotyper device is separated by semi-colons and has empty columns which prevent
an easy import by R. This function reads the report correctly as a tibble.
}
\details{
The headerless table contains identification information for each target processed by
the Biotyper device:
\itemize{
\item \code{bruker_quality}: a character encoding the quality of the identification with potentially multiple "+" symbol or only one "-"
\item \code{bruker_species}: the species name associated with the MALDI spectrum analyzed.
\item \code{bruker_taxid}: the NCBI Taxonomy Identifier of the species name in the column species
\item \code{bruker_hash}: a hash from an undocumented checksum function probably to encode the database entry.
\item \code{bruker_log}: the log-score of the identification.
}
}
\examples{
# Example with your dataset in "inst/"
biotyper <- system.file("biotyper.csv", package = "maldipickr")
read_biotyper_report(biotyper)
}
