% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_import_and_process_spectra.R
\name{tar_import_and_process_spectra}
\alias{tar_import_and_process_spectra}
\title{Import and process checked spectra using targets}
\usage{
tar_import_and_process_spectra(
  name,
  raw_spectra_directories,
  tolerance,
  format = targets::tar_option_get("format")
)
}
\arguments{
\item{name}{A character indicating the prefix of all targets created by the factory.
For instance, \code{name = anaerobe} will create the target \code{anaerobe_spectra_raw} among others.}

\item{raw_spectra_directories}{A vector of paths to directories containing MALDI Biotyper spectra files. This is similar to the \code{biotyper_directory} parameter from \link{import_biotyper_spectra}, but as a character vector.}

\item{tolerance}{A numeric indicating the accepted tolerance to the spectra length.
The default value is the machine numerical precision and is close to 1.5e-8.}

\item{format}{Optional storage format for the target's return value.
With the exception of \code{format = "file"}, each target
gets a file in \verb{_targets/objects}, and each format is a different
way to save and load this file. See the "Storage formats" section
for a detailed list of possible data storage formats.}
}
\value{
A target object
}
\description{
Given a vector of paths to \code{acqus} and \code{acqu} MALDI Biotyper directories, this targets
factory facilitates the steps from raw spectra to quality-checked processed spectra.
}
\examples{
if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
  targets::tar_dir({ # tar_dir() runs code from a temporary directory.
    targets::tar_script({
        library(maldipickr)
        list(
          tar_spectra_import(
            name = "example",
            raw_spectra_directories = system.file(
              "toy-species-spectra",
              package = "maldipickr"),
            tolerance = 1
          )
        )},ask = FALSE)
    targets::tar_make()
  })
}
}
