% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_pick_with_similarity.R
\name{tar_pick_with_similarity}
\alias{tar_pick_with_similarity}
\title{Delineate clusters of spectra to be picked using targets}
\usage{
tar_pick_with_similarity(name, targets_spectra, threshold, ...)
}
\arguments{
\item{name}{A symbol indicating the prefix of all targets created by the factory.
For instance, calling \code{tar_pick_with_similarity(anaerobe, ...)} will create
the target \code{anaerobe_sim_interpolated} among others (see the Value section).}

\item{targets_spectra}{A list of targets produced by \link{tar_import_and_process_spectra} that should contains one or more targets named \verb{*_processed}.}

\item{threshold}{A numeric value indicating the minimal cosine similarity between two spectra.}

\item{...}{Arguments passed to \link{pick_spectra}}
}
\value{
A list of target objects whose names use the \code{name} argument as a prefix:
\itemize{
\item \verb{*_fm_interpolated} (e.g.,  \code{anaerobe_fm_interpolated}): a matrix produced by \link{merge_processed_spectra}.
\item \verb{*_sim_interpolated} (e.g.,  \code{anaerobe_sim_interpolated}): a symetric cosine similarity matrix produced by \link[coop:cosine]{coop::tcosine}.
}
}
\description{
Description
}
\examples{
tar_pick_with_similarity()
}
