% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldipickr-package.R
\docType{package}
\name{maldipickr-package}
\alias{maldipickr}
\alias{maldipickr-package}
\title{maldipickr: Dereplicate And Cherry-Pick Mass Spectrometry Spectra}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convenient wrapper functions for the analysis of matrix-assisted laser desorption/ionization-time-of-flight (MALDI-TOF) spectra data in order to select only representative spectra (also called cherry-pick). The package covers the preprocessing and dereplication steps needed to cluster MALDI-TOF spectra before the final cherry-picking step. It enables the easy exclusion of spectra and/or clusters to accommodate complex cherry-picking strategies. Alternatively, cherry-picking using taxonomic identification MALDI-TOF data is made easy with functions to import inconsistently formatted reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClavelLab/maldipickr}
  \item Report bugs at \url{https://github.com/ClavelLab/maldipickr/issues}
}

}
\author{
\strong{Maintainer}: Charlie Pauvert \email{cpauvert@ukaachen.de} (\href{https://orcid.org/0000-0001-9832-2507}{ORCID})

Other contributors:
\itemize{
  \item David Wylensek (\href{https://orcid.org/0000-0002-8424-5712}{ORCID}) [contributor]
  \item Selina Nüchtern [contributor]
  \item Thomas Clavel (\href{https://orcid.org/0000-0002-7229-5595}{ORCID}) [contributor, funder]
}

}
\keyword{internal}
