% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectra.R
\name{process_spectra}
\alias{process_spectra}
\title{Process Bruker MALDI Biotyper spectra à la Strejcek et al. (2018)}
\usage{
process_spectra(spectra_list, rds_prefix = NULL)
}
\arguments{
\item{spectra_list}{A list of \link[MALDIquant:MassSpectrum-class]{MALDIquant::MassSpectrum} objects.}

\item{rds_prefix}{A character indicating the prefix for the \code{.RDS} output files to be written in the \code{processed} directory. By default, no prefix are given and thus no files are written.}
}
\value{
A named list of three objects:
\itemize{
\item \code{spectra}: a list the length of the spectra list of \link[MALDIquant:MassSpectrum-class]{MALDIquant::MassSpectrum} objects.
\item \code{peaks}: a list the length of the spectra list of \link[MALDIquant:MassPeaks-class]{MALDIquant::MassPeaks} objects.
\item \code{metadata}: a data frame indicating the median signal-to-noise ratio and peaks number for all spectra list, with their names as rownames.
}
}
\description{
Process Bruker MALDI Biotyper spectra à la Strejcek et al. (2018)
}
\details{
Based on the original implementation, the function performs the following tasks:
\enumerate{
\item Square-root transformation
\item Mass range trimming to 4-10 kDA
\item Signal smoothing
\item Baseline Correction
\item Normalisation
\item Peak detection
6.1 A peak filtering step has been added to discard peaks with a negative signal-to-noise ratio probably due to being on the edge of the mass range.
}
}
\note{
The original R code on which this function is based is accessible at: \url{https://github.com/strejcem/MALDIvs16S}
}
\examples{
directory_for_biotyper_spectra <- system.file("toy-species-spectra", package = "maldipickr")
spectra_list <- import_biotyper_spectra(directory_for_biotyper_spectra)
process_spectra(spectra_list)
}
\references{
Strejcek, Michal, Tereza Smrhova, Petra Junkova, and Ondrej Uhlik. “Whole-Cell MALDI-TOF MS versus 16S Rrna Gene Analysis for Identification and Dereplication of Recurrent Bacterial Isolates.” \emph{Frontiers in Microbiology} 9 (2018). \url{https://doi.org/10.3389/fmicb.2018.01294}.
}
\seealso{
\link{import_biotyper_spectra}
}
