% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_pick_with_similarity.R
\name{tar_pick_with_similarity}
\alias{tar_pick_with_similarity}
\title{Delineate clusters of spectra to be picked using targets}
\usage{
tar_pick_with_similarity(name, targets_spectra, threshold, ...)
}
\arguments{
\item{name}{A symbol indicating the prefix of all targets created by the factory.
For instance, calling \code{tar_pick_with_similarity(anaerobe, ...)} will create
the target \code{anaerobe_sim_interpolated} among others (see the Value section).}

\item{targets_spectra}{A list of targets produced by \link{tar_import_and_process_spectra} that should contains one or more targets named \verb{*_processed}.}

\item{threshold}{A numeric value indicating the minimal cosine similarity between two spectra.}

\item{...}{Arguments passed to \link{pick_spectra}}
}
\value{
A list of target objects whose names use the \code{name} argument as a prefix:
\itemize{
\item \verb{*_fm_interpolated} (e.g.,  \code{anaerobe_fm_interpolated}): a matrix produced by \link{merge_processed_spectra}.
\item \verb{*_sim_interpolated} (e.g.,  \code{anaerobe_sim_interpolated}): a symetric cosine similarity matrix produced by \link[coop:cosine]{coop::tcosine}.
\item \verb{*_df_interpolated} (e.g.,  \code{anaerobe_df_interpolated}): a tibble with the membership (i.e., which cluster label) each spectra belongs to produced by \link{delineate_with_similarity}.
\item \verb{*_processed_metadata} (e.g.,  \code{anaerobe_processed_metadata}): a tibble of aggregated technical metadata for each spectra.
\item \verb{*_clusters} (e.g.,  \code{anaerobe_clusters}): a tibble indicating with the previous metadata and which spectra was chosen as reference produced by \link{set_reference_spectra}.
\item \verb{*_picked} (e.g.,  \code{anaerobe_picked}): a tibble containing all the previous metadata but more importantly which spectra should be picked produced by \link{pick_spectra}.
}
}
\description{
Given upstream targets of processed spectra (from \link{tar_import_and_process_spectra})
this target factory facilitates the steps from quality-checked
processed spectra to clusters of spectra. See \link[targets:tar_target]{targets::tar_target} for more
information about what are target objects.
}
\examples{
\dontrun{tar_pick_with_similarity()}
}
